refactor : ♻️  회원가입 이동 텍스트 컴포넌트 분리
